requirements.txt
numpy
nibabel
scipy
matplotlib
scikit-image
scikit-learn
opencv-contrib-python
numpy-stl
cython
beautifulsoup4
lxml
# python-pcl